Clean and optimize code
I’ve moved most of the HTML code to a separate, static HTML file
(report.html). The file is cleared, the static content added, then the
dynamic content is replaced. This removes a lot of unnecessary Python
code. I’ve also tidied up the code to make it more readable and
efficient.